add wallet-signTransaction.test.ts
