--- test/_fixtures/signAndReturnTransactionSuccess.json
@@ -1,6 +1,0 @@
-{
-  "signerAccountId": "0.0.12345",
-  "transaction": {
-    "bytes": "CqYBKqMBCjgKGQoMCJ7C1aYGEMixkZsBEgcIABAAGLlgGAASBggAEAAYAxiAhK9fIgIIeDIAwgEHMgUIgM7aAxJnCmUKIQJ4J53yGuPNMGEGJ7HkI+u3QFxUuAOa9VLEtFj7Y6qNMzJAj/sYjPtrDkNcfIc/bRgBABLgFhwlUx2/1bDQqgyF40yxhxRPupYB1VRx/Tr5tNdVvJ3H3Ifn1Wy475jsfeL2/A=="
-  }
-}
